Use msgpack cache file for faster runs

+++ src/decomp_data.rs
@@ -71,6 +71,15 @@ impl DecompData {
                 .success());
         }
 
+        let cache_file_path = repo.join("sm64gs2pc.msgpack");
+
+        if cache_file_path.exists() {
+            return rmp_serde::decode::from_read(BufReader::new(
+                File::open(cache_file_path).unwrap(),
+            ))
+            .unwrap();
+        }
+
         std::fs::copy(base_rom, repo.join("baserom.us.z64")).unwrap();
 
         assert!(Command::new("make")
@@ -203,6 +212,9 @@ impl DecompData {
                 decomp_data.structs.insert(decl.name, struct_);
             }
         }
+
+        rmp_serde::encode::write(&mut File::create(cache_file_path).unwrap(), &decomp_data)
+            .unwrap();
 
         decomp_data
     }
